+++ frontend/src/metabase/admin/datamodel/components/AggregationPopover/AggregationPopover.styled.tsx
@@ -1,7 +1,7 @@
 import styled from "@emotion/styled";
 import AccordionList from "metabase/core/components/AccordionList";
 import { color } from "metabase/lib/colors";
-import FieldList from "../../../../query_builder/components/FieldList";
+import { FieldList } from "../FieldList";
 
 export const AggregationItemList = styled(AccordionList)`
   color: ${color("summarize")};
