+++ frontend/src/metabase/admin/datamodel/components/AggregationWidget/AggregationWidget.jsx
@@ -5,12 +5,12 @@ import { t } from "ttag";
 
 import ControlledPopoverWithTrigger from "metabase/components/PopoverWithTrigger/ControlledPopoverWithTrigger";
 import { isRows } from "metabase-lib/queries/utils/aggregation";
-import AggregationPopover from "../../admin/datamodel/components/AggregationPopover";
-import Clearable from "./Clearable";
+import AggregationPopover from "../AggregationPopover";
+import Clearable from "../../../../query_builder/components/Clearable";
 import { AggregationLabel } from "./AggregationWidget.styled";
 // NOTE: lots of duplication between AggregationWidget and BreakoutWidget
 
-export default class AggregationWidget extends Component {
+export class AggregationWidget extends Component {
   constructor(props, context) {
     super(props, context);
