+++ frontend/src/metabase/admin/datamodel/components/AggregationWidget/index.ts
@@ -1,0 +1,1 @@
+export * from "./AggregationWidget";
