+++ frontend/src/metabase/admin/datamodel/components/BreakoutPopover/BreakoutPopover.styled.tsx
@@ -1,6 +1,6 @@
 import styled from "@emotion/styled";
 import { color } from "metabase/lib/colors";
-import FieldList from "metabase/query_builder/components/FieldList";
+import { FieldList } from "../FieldList";
 
 export const BreakoutFieldList = styled(FieldList)`
   color: ${color("summarize")};
