+++ frontend/src/metabase/admin/datamodel/components/BreakoutPopover/index.ts
@@ -1,0 +1,1 @@
+export * from "./BreakoutPopover";
