+++ frontend/src/metabase/admin/datamodel/components/BreakoutWidget/index.ts
@@ -1,0 +1,1 @@
+export * from "./BreakoutWidget";
