+++ frontend/src/metabase/admin/datamodel/components/DimensionList/index.ts
@@ -1,0 +1,1 @@
+export * from "./DimensionList";
