+++ frontend/src/metabase/admin/datamodel/components/DimensionPicker/index.ts
@@ -1,0 +1,1 @@
+export * from "./DimensionPicker";
