+++ frontend/src/metabase/admin/datamodel/components/FieldList/index.ts
@@ -1,0 +1,1 @@
+export * from "./FieldList";
