+++ frontend/src/metabase/admin/datamodel/components/FilterWidget/FilterWidget.tsx
@@ -7,7 +7,7 @@ import FilterComponent from "metabase/query_builder/components/Filter";
 
 import type Filter from "metabase-lib/queries/structured/Filter";
 import type StructuredQuery from "metabase-lib/queries/StructuredQuery";
-import { FilterPopover } from "./FilterPopover";
+import { FilterPopover } from "../FilterPopover";
 import {
   FilterField,
   FilterOperator,
@@ -70,8 +70,7 @@ type State = {
   isOpen: boolean;
 };
 
-// eslint-disable-next-line import/no-default-export -- deprecated usage
-export default class FilterWidget extends Component<Props, State> {
+export class FilterWidget extends Component<Props, State> {
   rootRef: React.RefObject<HTMLDivElement>;
 
   constructor(props: Props) {
