+++ frontend/src/metabase/admin/datamodel/components/FilterWidget/index.ts
@@ -1,0 +1,1 @@
+export * from "./FilterWidget";
