+++ frontend/src/metabase/admin/datamodel/components/FilterWidgetList/index.ts
@@ -1,0 +1,1 @@
+export * from "./FilterWidgetList";
