+++ frontend/src/metabase/admin/datamodel/components/GuiQueryEditor/index.ts
@@ -1,0 +1,1 @@
+export * from "./GuiQueryEditor";
